requirements.txt
betfair.py==0.2.2
decorator==4.0.4
enum34==1.0.4
Flask==0.10.1
inflection==0.3.1
invoke==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mock==1.0.1
python-dateutil==2.4.2
requests==2.8.1
schematics==1.1.0
six==1.10.0
urllib3==1.12
Werkzeug==0.10.4
wheel==0.24.0
